add result pattern

+++ src/Auction.Application/Lots/Commands/DeleteLot/DeleteLot.cs
@@ -1,14 +1,15 @@
 ﻿using Auction.Application.Common;
 using Auction.Application.Common.Abstractions.UnitOfWork;
+using Auction.Application.Common.Models;
 using Auction.Domain.Entities;
 using Quartz;
 
 namespace Auction.Application.Lots.Commands.DeleteLot;
 
 [Authorize(Role.Moderator | Role.Admin)]
-public record DeleteLotCommand(Guid LotId) : IRequest<bool>;
+public record DeleteLotCommand(Guid LotId) : IRequest<Result>;
 
-public class DeleteLotCommandHandler : IRequestHandler<DeleteLotCommand, bool>
+public class DeleteLotCommandHandler : IRequestHandler<DeleteLotCommand, Result>
 {
     private readonly IUnitOfWork _unitOfWork;
     private readonly IScheduler _scheduler;
@@ -21,14 +22,21 @@ public class DeleteLotCommandHandler : IRequestHandler<DeleteLotCommand, bool>
         _scheduler = scheduler;
     }
 
-    public async Task<bool> Handle(DeleteLotCommand command, CancellationToken cancellationToken)
+    public async Task<Result> Handle(DeleteLotCommand command, CancellationToken cancellationToken)
     {
         using var connection = _unitOfWork.Create();
 
+        var lot = await connection.Repositories.LotRepository.GetById(command.LotId);
+
+        if (lot is null)
+        {
+            return Result.Failure(LotError.NotFound);
+        }
+
         await connection.Repositories.LotRepository.Delete(command.LotId);
 
-        connection.SaveChanges();
+        await connection.SaveChangesAsync();
 
-        return true;
+        return Result.Success();
     }
 }
